+++ avatars/__init__.py
@@ -1,2 +1,2 @@
 # Also modify in pyproject.toml
-__version__ = "0.8.0"
+__version__ = "0.10.0"