+++ avatars/client.py
@@ -1,5 +1,5 @@
 # This file has been generated - DO NOT MODIFY
-# API Version : 3.0.1-739cf4fbfd2c425a13dd912b8f3d8e873174c36d
+# API Version : 5.1.0-5d5865f0a4b03a10f73a0894938c6843af952096
 
 
 import warnings